+++ .meta-updater/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm-private/updater
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/lockfile-file@8.1.3
+
 ## 1.0.5
 
 ### Patch Changes