+++ exec/prepare-package/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/prepare-package
 
+## 5.0.13
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/lifecycle@16.0.1
+  - @pnpm/read-package-json@8.0.4
+
 ## 5.0.12
 
 ### Patch Changes