+++ fetching/directory-fetcher/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/directory-fetcher
 
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/fetcher-base@15.0.2
+  - @pnpm/read-project-manifest@5.0.7
+  - @pnpm/resolver-base@10.0.3
+
 ## 7.0.1
 
 ### Patch Changes