+++ fetching/git-fetcher/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @pnpm/git-fetcher
 
+## 12.0.5
+
+### Patch Changes
+
+- Updated dependencies [08b65ff78]
+  - @pnpm/worker@0.3.4
+  - @pnpm/prepare-package@5.0.13
+
+## 12.0.4
+
+### Patch Changes
+
+- @pnpm/worker@0.3.3
+
+## 12.0.3
+
+### Patch Changes
+
+- @pnpm/worker@0.3.2
+
+## 12.0.2
+
+### Patch Changes
+
+- @pnpm/prepare-package@5.0.13
+- @pnpm/fetcher-base@15.0.2
+- @pnpm/worker@0.3.1
+
 ## 12.0.1
 
 ### Patch Changes