+++ lockfile/lockfile-types/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/lockfile-types
 
+## 5.1.2
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+
 ## 5.1.1
 
 ### Patch Changes