+++ packages/dependency-path/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/dependency-path
 
+## 2.1.4
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/crypto.base32-hash@2.0.0
+
 ## 2.1.3
 
 ### Patch Changes