+++ pkg-manager/package-bins/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/package-bins
 
+## 8.0.3
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+
 ## 8.0.2
 
 ### Patch Changes