+++ pkg-manager/package-requester/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @pnpm/package-requester
 
+## 24.0.5
+
+### Patch Changes
+
+- 01bc58e2c: Update ssri to v10.0.5.
+- Updated dependencies [01bc58e2c]
+- Updated dependencies [08b65ff78]
+  - @pnpm/store.cafs@2.0.6
+  - @pnpm/worker@0.3.4
+
+## 24.0.4
+
+### Patch Changes
+
+- @pnpm/worker@0.3.3
+
+## 24.0.3
+
+### Patch Changes
+
+- @pnpm/worker@0.3.2
+
+## 24.0.2
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/package-is-installable@8.0.5
+  - @pnpm/fetcher-base@15.0.2
+  - @pnpm/core-loggers@9.0.3
+  - @pnpm/dependency-path@2.1.4
+  - @pnpm/read-package-json@8.0.4
+  - @pnpm/resolver-base@10.0.3
+  - @pnpm/store.cafs@2.0.5
+  - @pnpm/store-controller-types@17.0.1
+  - @pnpm/worker@0.3.1
+  - @pnpm/pick-fetcher@2.0.1
+
 ## 24.0.1
 
 ### Patch Changes