+++ pkg-manager/remove-bins/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/remove-bins
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/core-loggers@9.0.3
+  - @pnpm/package-bins@8.0.3
+  - @pnpm/read-package-json@8.0.4
+
 ## 5.0.3
 
 ### Patch Changes