+++ pkg-manifest/manifest-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/manifest-utils
 
+## 5.0.4
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/core-loggers@9.0.3
+
 ## 5.0.3
 
 ### Patch Changes