+++ pkg-manifest/read-project-manifest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/read-project-manifest
 
+## 5.0.7
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/write-project-manifest@5.0.3
+
 ## 5.0.6
 
 ### Patch Changes