+++ store/cafs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pnpm/store.cafs
 
+## 2.0.6
+
+### Patch Changes
+
+- 01bc58e2c: Update ssri to v10.0.5.
+
+## 2.0.5
+
+### Patch Changes
+
+- @pnpm/fetcher-base@15.0.2
+- @pnpm/store-controller-types@17.0.1
+
 ## 2.0.4
 
 ### Patch Changes