+++ store/server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/server
 
+## 17.0.1
+
+### Patch Changes
+
+- Updated dependencies [d774a3196]
+  - @pnpm/types@9.3.0
+  - @pnpm/store-controller-types@17.0.1
+  - @pnpm/fetch@7.0.4
+
 ## 17.0.0
 
 ### Major Changes